requirements.txt
fastapi[standard]~=0.116.1
sqlmodel~=0.0.24
SQLAlchemy~=2.0.41
psycopg2-binary~=2.9.10
pydantic~=2.11.7
alembic~=1.16.4
celery~=5.5.3
redis~=6.2.0
diskcache~=5.6.3
requests~=2.32.4
urllib3~=2.5.0
websockets~=15.0.1
broadcaster~=0.3.1
pyjwt~=2.10.1
bcrypt~=4.3.0
email-validator~=2.2.0
flower~=2.0.1
transliterate~=1.10.2
beautifulsoup4~=4.14.3
elasticsearch~=9.3.0
aiohttp~=3.13.3